Phabricator: Add an initial main.
This will be used to run the Webserver and the gRPC services.
Also add a Readme in /src/native, which describes it's purpose.

See the commented code in the file.
And lastly add a .clang-format file based on the Phabricator PHP-Style.
The Style is based on Google's C++ with some additions.